Fix the ground-term detection of jj-coupled ions and of unresolved splittings
Code review of the branch found that a single interval-rule tolerance cannot
serve both comparisons the loop makes. Comparing a splitting with the one below
it stays inside the multiplet, where the rule holds to a factor of about three.
Comparing with the splitting above happens at the lowest level, where for a
single-level ground term the reference is the fine structure of the next term:
those boundaries need a factor of 25 or more, while the widest real multiplet
needs 3.4. With one tolerance of 3 the C-like 3P of Fe XXI and the S-like 3P of
Fe XI, both in the shipped classic dataset, collapsed to a single level; Fe XXI
was a regression against the code this branch replaces. Splitting the tolerance
in two fixes both and changes nothing else in any dataset.

Also, from the same review:

- The reference splitting is now the nearest one below that the data resolves as
  non-zero, so a splitting rounded to zero no longer takes the whole term with
  it. When neither neighbour is usable the unresolved splitting below is used
  anyway, so that a large jump after a degenerate pair still ends the term
  instead of being accepted with no energy test at all.
- splittings_consistent() fails closed on a negative splitting (levels out of
  energy order) and on a reference pair with no J step, rather than returning a
  vacuous true from a zero factor on one side.
- The parity test rounds the ground level's statistical weight instead of
  truncating it, so a weight written as 6.999999 still tests as odd.
- Mismatched span lengths are asserted rather than silently truncated.
- The statistical weight progression is checked against a direction computed
  once, dropping the mutable state and the guard it needed.
- Unit tests cover the highly-charged regime, which they did not before.

+++ input.cc
@@ -690,8 +690,8 @@ void add_transitions_to_unsorted_linelist(const int element, const int ion,
 }
 
 auto calculate_nlevels_groundterm(const int element, const int ion) -> int {
-  const auto nlevels = static_cast<size_t>(get_nlevels(element, ion));
-  const auto levelstart = static_cast<size_t>(get_ionuniquelevelindexstart(element, ion));
+  const auto nlevels = get_nlevels(element, ion);
+  const auto levelstart = get_ionuniquelevelindexstart(element, ion);
   return count_groundterm_levels(globals::alllevels.epsilon.subspan(levelstart, nlevels),
                                  globals::alllevels.statweight.subspan(levelstart, nlevels));
 }
